Bugfix and cleanup Info on source in UI

+++ MediaPlayer.h
@@ -299,7 +299,7 @@ private:
 
         GstClockTime last_time;
         GstClockTime tic_time;
-        int nbFrames;
+        long nbFrames;
         gdouble fps;
     public:
         TimeCounter();
